[feature] Handling for multiple insertion of same resource added
A handling for the insertion of the same resource more than once has
been added. Such entries are highlighted with a slighly different color
and, when the user pushed the "add" button, he will be told that the
entry he wants to add has already been added. The user can decide if
he/she wants to continue or stop the insertion.